refactor: move tests to separate tests directory
- Move all test files from thetagang/ to tests/
- Add tests/__init__.py
- Fix import in test_util.py to use relative import
- Add CLAUDE.md with development guidance
