+++ tests/test_trades.py
